Hash tables must also resize if the max probe distance is hit on an empty slot.
Not just for the case of max probe distance being hit during the move
operation to open up a slot to maintain the Robin Hood invariant.

This is a bug, but it's extremely unlikely that anyone would hit it. It would
take either extremely pathological data, or an extremely large data set with
merely regular pathological data :-)

For reference, storing all the individually named Unicode code points
doesn't exceed a probe distance of 30, whilst this would only hit if one
both had a probe distance for the new insert of 255, *and* also no other
entries at lower probe distances that needed moving.