Update cur_items as each entry is (re)added in hash_insert_internal().
Set it to 0 in hash_allocate_common(), instead of setting it to 0 in
MVP_hash_initial_allocate(), and to the previous control structure's value
of cur_items in maybe_grow_hash().

This *might* cause more write cache activity, but it simplifies the logic
considerably, and means that the hash is always in a "legal" state during
the grow operation, meaning that it's now viable to call hash_fsck_internal
on it.