Split apart `MVM_str_hash_kompromat` and `MVM_str_hash_allocated_items`.
The former is the count of the entries currently in use. The latter is the
count of entries allocated. They have been the same until now, but they will
soon differ.

Attempt to make various local variable names consistent with the functions
whose return values they hold. Rename `hash_true_size` to
`calc_entries_in_use` as this is clear about which size we're talking about.
(The size of entries in use for this max probe distance, not the total size
allocated, which corresponds to the max probe distance limit.)