Fix an off-by-one in how we were storing probe distances in hash tables.
When we start out with an allocation of 8 + 5 buckets, max probe distance is
*six*. We store the sixth value at metadata location ideal_bucket + 5

We had been storing max_probe_distance and limit off by one, but had never
encountered the problem because it would only matter for pathological data
in large hashtables, where we needed a probe distance of 254.

Also fix an off-by-one in an assertion in the ...insert routines. For insert,
the loop must not iterate to any probe distance greater than the (current)
maximum probe distance. For fetch and delete, the loop is permitted to
reach (and read) one beyond the maximum probe distance - but if so, it will
always read a probe distance which is lower than the current probe distance,
and hence hit "not found" and terminate the loop. This is how the loop
terminates when the max probe distance is reached without needing an explicit
test for it, and why we need an initialised sentinel byte at the end of the
metadata.

Remove a comment from `MVM_str_hash_delete_nocheck` which is incorrect.
This was copied from `MVM_ptr_hash_fetch_and_delete`, where what it describes
is correct.