Testing with all metadata hash bits set to 1 reveals a bug in delete.
When moving existing entries down to close the gap caused by the deleted
item, we need to compare with the "metadata increment" plus all bits set.
Otherwise we can move some entries which are actually at probe distance 1
(ie the ideal bucket) which is clearly a bug. (fsck spots this.)