Store some of each entry's hash in the low bits of the metadata.
This permits fast rejection of many "cannot match" cases without needing any
extra CPU work.

The number of bits of hash stored is calculated dynamically. We commence
with 3 bits for probe distance and 5 for hash, and shift more bits to
probe distance whenever we would otherwise overflow the space allocated for
probe distance. Worst case we end up with 8 bits for probe distance and 0 for
hashes, but only when we need to store probe distances >= 127.

Usually we double the size of hash table before things get this bad. We reset
to 3 bits/5 bits on each hashtable resize.