Fix interning of parametric types in precompilation
When the same parametric type gets created by multiple, different compilation
units, we need to unify them, when loading those units. Otherwise a Foo[Bar]
created in one unit would not match a Foo[Bar] created in another. This
unification is called interning.

Three separate problems prevented this from working as intended. First, we only
wrote paremtric types to the parametric intern table, if they were created
outside the current compilation on the premise that there's nothing to intern
when a type originated in the compilation unit we're loading it from. But a
trivial example of two separate modules creating parametrics with both modules
getting loaded into the same program refutes this premise. Instead, we really
should write all parametric types to the table.

Secondly parametrics can be nested: Foo[Bar[Baz]] or Foo[Bar:D]. In that
example if we try to find Foo[Bar[Baz]] before Bar[Baz] is resolved, we won't
find it, as we're looking for the wrong Bar[Baz]. We get around this by doing
passes through the table until we cannot find any new matches.

And thirdly the sequencing was wrong. We were resolving parametrics before
processing repossessions. Since the former requires us to deserialize the
involved types, we could end up deserializing a not-yet-repossessed type.