For the pthread_setname_np probe, use an auto char array instead of malloc.
There's no requirement from the API of pthread_setname_np to pass in a
pointer from `malloc`. However, as the probe had been implemented, it would
fail under an ASAN build, thus Configure.pl would incorrectly assume that
pthread_setname_np was not available. because ASAN defaults to leak checking
on exit, the leak checking is aggressive and sets a non-zero exit code on
leaks, and `malloc` without `free` is a leak.

Hence use a char array, and avoid `malloc`. This way the probe also passes
under ASAN, and we get to test the relevant runtime code with ASAN too.