Report the lines that caused coverage
This makes the program even more specific to parsing a coverage report
of a Rakudo spectest. But the benefit is you can see what lines in the
tests cover lines in Rakudo. It would probably be better to create a
table, or hoverover/popup (maybe on the line number?), but I don't know
HTML/CSS/JS/etc, so just sticking them at the end of the line.