The inline `memcpy` replacement can be used on all little-endian platforms.
The `memcpy` in `MVM_serialization_read_int` was to avoid unaligned access
to a MVMint64 value. The replacement code that unwraps and inlines it only
does byte access to memory, so it's quite safe to *also* use this on
platforms that can't do unaligned 64-bit reads.

(If the compiler on platforms that can read unaligned spots that it can
optimise our byte accesses to larger reads, that's just fine, because it
doesn't affect the platforms where this is fatal.)