Add casts to calls to `switch_endian` to avoid compiler warnings.
`switch_endian` takes a `char *` pointer - for the read* functions where we
are passing a pointer to the integer variable that we read into, we need a
cast else the compiler rightly warns.

Mark `switch_endian` itself as MVM_STATIC_INLINE - the compiler probably
assumes this already, but let's just be clear.