Fix zeroing of reallocated memory in MVM_recalloc
This function assumes that the new size of memory to be reallocated is
greater than or equal to its old size. Should it be lesser instead, this
will attempt to zero memory that is out of bounds, leading to a
segfault.  Only attempt to zero reallocated memory if its size has
grown.