In CStruct `bind_attribute`, don't `set_ptr_at_offset` for inlined CArray.
`bind_attribute` had been calling `set_ptr_at_offset` for all `CArray`s,
whether inlined or not. That writes a pointer at the slot offset.

For the corner case of the inlined array being both the last member in the
structure and smaller than the size of a pointer
(eg `HAS int32 @.b[1] is CArray;` on a 64 bit system),
this causes a pointer write beyond the end of the allocated storage.
More generally, on a system which faults unaligned pointer writes, this
write faults in the more realistic case where the previous structure
member(s) don't add up to pointer-sized alignment.

This is clearly a bug.

It seems that the write to the slot is not needed in this case. This commit
takes the same approach as commit e4ff6949548e4f18 from May 2020:
    Setup inlined CStruct assigned during construction

    Fixes https://github.com/rakudo/rakudo/issues/3687

by `break`ing out of the switch statement for the inlined case.

The bug itself seems to date back to the code first being added in Aug 2018
by commit ac3d3c76954fa3c1:
    Fix alignment/mem issue of inlined CArrays

    Now CArrays that are inlined in CStructs are handled correctly
    regarding alignment of the CArray within the struct. Also accessing
    the slots of the inlined array works.