MVM_nativeref_{read,write}_lex_i should handle uint8, uint16, uint32
Previously these three (and uint64) all took the `default:` path through
the case statement, which meant that they were read/written as 64 bits, eg:

        default:
            return var->i64;

This works fine on little endian systems, where the least significant byte
of the 64 bit value is at the the same memory address as the 8 bit value,
but on big endian systems it all goes wrong.

The two routines already had cases for the various sized signed types.
We need case statements to cover the 3 smaller unsigned types.

For clarity, I think it better to cover all 4 unsigned types explicitly,
and use the corresponding members in the union. Likely the compiler will
fold the signed and unsigned cases together in the final code, but I'd
prefer it to be doing that, rather than us making assumptions about
equivalence.