No need to MVMROOT `root` twice in `get_attribute`.
This inefficiency was introduced 6 weeks ago as a side effect of commit
b7151da7d7016c07:
    Fix possible memory corruption in CStruct's and CPPStruct's get_attribute

    An untimely GC run may move *root before we get to assign the newly created
    object into the child_objs slot. This could cause us to over write unrelated
    memory. Fix by MVMROOTing root.
    Same fix as commit d76ebbe49ada2d4649b9237db86fd36e004d6612 did for CUnion

which wrapped a large block of code in MVMROOT to root `root`, not realising
that `root` was already rooted in this inner MVMROOT2.

*This* error didn't exist in CPPStruct.c because the MVMROOT2 was added only
to CStruct.c in July 2019 by commit 4a90335f322c7395:
    Fix possible access to fromspace in CStruct's get_attribute

    root could be moved by a GC run triggered by string decoding. Since CStruct's
    body is inlined into MVMCStruct the body would be moved along, so we must get
    a fresh pointer to that, too.

a bugfix that logically belonged also in CPPStruct.c (and effectively was
addressed by commit b7151da7d7016c07)

The code in CPPStruct.c and CStruct.c *should* be pretty much identical, but
the DRY violation means that the two have diverged as one gets bug fixes but
not the other. I'm loathe to fix that *right* now - better to address the
current known bugs in the well-tested CStruct.c, and then address the bigger
problem of more effectively maintaining two necessarily *near* identical
codebases.