Fix interning of nested parametric types
Parametrics can be nested like Hash[Array[Int:D]]. To be able to resolve a
parametric like that, the inner parts already have to be interned. Otherwise
the type checks in MVM_6model_parametric_try_find_parameterization will fail.
Furthermore the deserialized not-interned objects and stables can stay
referenced, e.g. as a P6Opaque's attribute. This would cause issues later on.

To fix both issues we must not deserialize those not-interned objects right
away but instead, intern them on sight. When we process a parametric's
parameters for interning and such a parameter is itself a parametric, we need
to resolve that before moving on.

As a preparation for this new approach, we create and populate lookup tables
keyed by the index of the type object into the SC's root_objects and the index
of the STable into the SC's root_stables respectively. The value is the offset
into the parametric intern table, so we can easily load the information about
the parameterized type and the parameters for any type or type object we
deserialize.

In MVM_serialization_demand_object and MVM_serialization_demand_stable we first
check the lookup tables if the type (object) we try to deserialize is a
candidate for interning. If it is, we load the information from the param
intern table and try to find a matching candidate. As loading the parameterized
type and its parameters will go through MVM_serialization_demand_object and/or
MVM_serialization_demand_stable again, we achieve full recursion and avoid
deserializing or even stubbing anything we're not going to use anyway.