Fix interning of parametrics missing most cases
We only wrote paremtric types to the parametric intern table, if they were
created outside the current compilation on the premise that there's nothing to
intern when a type originated in the compilation unit we're loading it from.

It is however quite possible and even likely for multiple compilation unit to
create the same parametric type independedly and we still need to consolidate
those into the same type objects and STables, so cross module type checks will
succeed.

So forgo the the filtering of parametrics and instead add all of them to the
param intern table so we get a chance of interning them when deserializing.