Add MVM_THREAD_LOCAL to declare variables with thread local storage
If the compiler provides this, it's likely faster than UV's API, which is
(basically) POSIX thread local storage, and requires a function call for
each "get". UV appears to rely on Win32 compilers supporting this. All
recent gcc and clang versions support it, and seemingly also the vendor
compilers on Solaris and AIX.

(Likely also VMS, given that it originated in the ia64 ABI, but I suspect
UV support for VMS won't happen any time soon.)

Initially, just use this if UV defined UV_THREAD_LOCAL, which seems only to
be the case on Win32.