Use `foo(void)` instead of just `foo()` for functions that take no arguments
Whilst the two are equivalent in C++, in C the latter is just a declaration
not a prototype - it specifies the return type but not the parameters,
meaning that the compiler can't actually check the parameters passed and
report errors. An explicit `void` makes it a prototype, declaring that the
function takes no arguments.

(More wonderfully, it seems that even on some fairly recent versions of gcc,
is seems a function *definition* `static int foo() { ... }` doesn't act as a
prototype, meaning that even incorrect calls made to it later in the same
source file aren't faulted. clang does not stay silent on these.)

Found with a combination of `-Werror=strict-prototypes` and `git grep`.
We can't actually build with that compiler flag, because some of our
dependencies have header files that contain the problematic constructions,
and trying to work around these to turn the flag on by default probably
only creates fragile "solutions" and likely is more cost than benefit.