Work to replace the intcache with an approach that doesn't leak "does".
The goal is to replace the current intcache with a fast(er) way to generate
all boxed integers. Start with a proof of concept in MVM_repr_box_int.

The intcache intends to speed up boxing of commonly used values (-1 to 15).
It does that by holding pre-generated boxed objects for each value, and
returning a pointer to that self-same object. As it's a constant, this all
works, right?

Problem is that there are *some* things you can still do to a constant. So
for a value in the cache:

    my int $a = 1;
    my int $b = 1;
    $a does role { method Str() { "foo" } };
    $b.Str.say

we get "foo", whilst for a value not in the cache:

    my int $a = 42;
    my int $b = 42;
    $a does role { method Str() { "foo" } };
    $b.Str.say

we get the right answer (42).