MVM_box_{int,num,str,uint} mostly duplicate MVM_repr_box_{int,num,str,uint}
Convert them to thin wrappers. We can't make them inline because the JIT
needs to take their addresses.

MVM_box_{int,num,str} were added as part of the JIT work in July 2014 by
commit 47f6848e782f8d0e:
    Add boxing operators

    For simplicity, this means I moved the conditional initialization
    of a container to a function in coerce.c. This should make the
    non-jitted times somewhat slower, however I feel that'd be an
    acceptable compromise.

The relevant inline code was extracted from the `case` statements in interp.c
and moved to functions, so that it could be shared with the JIT. What wasn't
noticed at the time was that the inline code was already almost a duplicate of
code in reprconv.c

The uint variants of both functions were added later.