Minimally exact bigint/bigint => num conversion, including rounding.
Given that Rakudo parses decimal constants such as 0.01 as `Rat`s (ie
represents them internally as (1 / 100), and persists with this exact
representation as long as possible), this effectively means that the MoarVM
operator that converts `Rat`s to `Num`s implements Rakudo's decimal =>
binary floating point conversion.

Decimal to binary floating point conversion is messy, inherently lossy (in
the general case), subject to errors if not done carefully and correctly
(subtle errors such as drift - failure to round trip), and potentially slow,
and can require considering a lot more digits than seems intuitive for
correct rounding in the extreme corner cases.

(It's something approaching 750 decimal digits for the absolute worst case
value, to determine the correct rounding direction where the true value is
very close to the boundary between the nearest two values representable in
IEEE binary64, but I can no longer find the citation with the guilty
numbers.)

The previous code in `MVM_bigint_div_num` calculated the integer division
with 64 extra bits of precision, then implicitly truncated when converting
to a double. This handled most cases correctly, but one of the spectests
failed on 32 bit systems. (arm, ppc32, x86 - so not a compiler or CPU bug).

Specifically, the problematic test case was taking these two expressions:

    say Num(0.777777777777777777777);
    say Num(0.7777777777777777777771)

and confirming that the floating point conversion of the second is >= to
that of the first. Obviously it should be, because it's a slightly larger
decimal constant, but performing this comparison acts as a stress test on
the decimal to binary floating point conversion code.

The way the conversion is implemented is by calling `MVM_bigint_div_num`
to evaluate 777777777777777777777 / 1000000000000000000000
(and similarly for the second), and as stated just above, actually
implementing this as the big int calculation
(777777777777777777777 * 18446744073709551616) / 1000000000000000000000
then converting that to a double, then dividing by 2.0**64

The two (very) big integers corresponding to the values above are

    16938500042286714625715486403227953144517
    271016000676587434011482627366019954410936

The double conversion is the cause of the immediate problems here.
On a "64 bit platform", libtommath uses 60 bit "digits" internally; on a
32 bit platform it uses 28 bit "digits". This determines the base that it's
using internally.

(I believe "64 bit platform" is actually one with 128 bit integer support)

It sets its accumulator to 0.0, then for each digit starting with the *most*
significant, it loops
   multiply the accumulator by the base
   convert the current digit to a double
   adds it to its accumulator

stopping when it has consumed more than 64 bits of bigint, or has run out
of digits. As described above, the code then divided by 2.0**64 to reach
the correct answer.

On a 64 bit platform, all works well. Our two examples split into 60 bit
digits as shown, and the double conversion truncates to 53 bit mantissa
as shown:

31c7  1c71c71c71c71b8  ...
                  ^ double conversion 0b0001 (round up here)
31c71  c71c71c71c71ba4  ...
                    ^ double conversion 0b1... (round down here)

                ^
                |
but for both, THIS nibble is still 0xc after the addition, which is key.
IEEE floating point arithmetic is performed as if it's infinite precision,
and then the result is rounded to fit, and this is the rounding point,
and for both cases it rounds up, hence:

result: 0x1.8e38e38e38e39p+133
        0x1.8e38e38e38e39p+137

ie      0.7777777777777778
        0.7777777777777778

However, on 32 bit platforms, we get different results. Our two examples
split into 28 bit digits as shown:

31c71c  71c71c7  1c71b86   ...
31c71c7  1c71c71  c71ba4e   ...

We have to loop over 3 (of the 5) bigint digits.

For the former, the loop is

31c71c
31c71c71c71c7
31c71c71c71c71c71b86
             ^ 0b000. but remainder is 0b...1 1100 ... so round up

result: 0x1.8e38e38e38e39p+105
ie      0.7777777777777778

for the latter, the loop is

31c71c7
31c71c71c71c71
             ^ 0b000. but remainder is 0b...1 exactly so round to even

Here, round to even is round down. OH NOES. That bit is zero.
Next time round the loop the intermediate infinite precision value is:

31c71c71c71c70c71ba4e
             ^ 0b000. remainder is 0b...0 1100 .. so round DOWN

result: 0x1.8e38e38e38e38p+109
ie      0.7777777777777777

and hence 0.7777777777777777777771 is rounded down.

What should the fix be? MOAR BITS?

The code is trying to ensure enough precision by scaling up:

    int bbits = mp_count_bits(ib)+64;

So what if we try 65 extra bits instead? For this case, both `Num`s give
0.7777777777777778. As do 66 and 67 extra bits. But for 68 extra bits,
Num(0.777777777777777777777) now gives 0.7777777777777777
while Num(0.7777777777777777777771) gives 0.7777777777777778

because now the bit patterns misalign with the double rounding in different
ways. The results happen to be in the correct order, but they differ
from the results we got before.

There is no winning move here. Whatever fixed choice we pick will be wrong
for some values.

(The looping accumulator-as-double algorithm is also broken with 60 bit
bigint digits, but only for certain pathological integers at least 154
bits long which happen to lie *very* close to the boundaries between two
values that IEEE doubles can represent. Akin to the 28 bit examples above,
pathological values would really need all 154 binary bits to determine the
correct rounding direction but with only the first 54, the 53rd or 54th bit
would be rounded (or truncated) one way when the intermediate value was
stored in the accumulator, so that the next time round the loop when the
full value was (re)constructed within the IEEE double addition, that bit was
not the same as the original bitint, and so rounding when the wrong
direction. But no-one has noticed this yet.)

Hence this commit calculates the exact extra precision needed to ensure that
the integer division generates the needed mantissa directly, and then
performs the necessary rounding before directly writing out a complete
floating point representation.

For decimal conversions, it is bit-for-bit consistent with David M. Gay's
`strtod` (which seems to be taken as the gold standard in conversions),
not just for "normal" numbers, but also values that overflow, underflow
(into subnormals, zero or negative zero), and handles all the rounding
correctly for every test value I throw at it. (ie the last bit of the
mantissa is "correct", where "correct" is consistent with `strtod`).

Worst case this code will increase the inputs to the integer division by
2**54 (ie 17 decimal digits), whereas the previous code would always use
2**64 (ie 20 decimal digits, but always). The rest of the operations are
logical operations and bitshifts on integers. It should be slightly faster,
but more importantly I believe it to be canonically correct.