P6opaque's `spesh` must handle P6bigint values inline, as some are too big.
P6opaque's `spesh` will attempt to take a concrete P6bigint known value and
convert it to a `MVM_OP_const_i64`. This doesn't work if big integer's value
is to large than an MVMint64.

Previously the code was not considering this corner case. It was calling
`MVM_repr_get_int`, which would throw an exception. Not only did this
cause *that* spesh run to abort, it seems that it would leave spesh in an
inconsistent state such that future specialisations could fail spuriously.

This was causing t/spec/S02-types/declare.rakudo.moar and
t/spec/S02-types/int-uint.rakudo.moar to fail with an exception when run
with MVM_SPESH_NODELAY=1, but in *this* code (run in an EVAL):

    my int64 $var = 0;

because a previous EVAL inside a try block had caught an out-of-range
assignment to an int64 variable.