Calling `MVM_exception_throw_adhoc` in the spesh worker should be an oops.
Similarly no exceptions should be thrown in the event loop thread.

Also report these two special threads in `MVM_oops`, and treat a NULL
thread context pointer in oops as a bug worthy of a panic.

(Previously it would have "panic"ed because it would have crash due to the
NULL pointer dereference. This way, we give slightly better diagnostics.)