Use simpler stdio calls in exceptions.c where possible.
fputc('\n', stderr) is simpler than fprintf(...) or fwrite(...)
Because `fputs` is not varargs, the caller has a simpler ABI than `fprintf`
on some architectures (eg x86_64).

(Well, that's the theory anyway. In practice, it seems that a new enough gcc
optimises simple cases of either into calls to `fwrite`.)