Simplify and inline `MVM_nativecall_find_thread_context`
Use `MVM_get_running_threads_context` to get the thread context directly,
instead of needing to take a mutex and iterate over the linked list of
threads.

As the function is only used in one place, and is now much smaller, convert
it to an inline function, which means moving it into the header file.