Improve performance of repeated `unshift`ing
Prior to this, we always created a fixed amount of extra space (8
elements) at the start of the array, regardless of array size. With
this change we start accounting for the number of elements in the
array. This will make no difference for steady state unshift/pop,
but for situations where we repeatedly unshift will bring the
performance in line with that of push. Fixes #1382.