Try to fix "Corrupt multi dispatch cache: cur_node != 0, re-check == 0"
MVM_multi_cache_find did a couple more checks when looking for a multi
candidate than MVM_multi_cache_add, i.e. it was more picky. This could lead to
the check for a pre-existing candidate to not find one and the code for finding
the right tree node to extend getting surprised by finding a matching candidate
after all.