Fix bogus "Corrupt multi dispatch cache" panic when arguments change
Before we can add a candidate to the multi dispatch cache, we need to find the
right node of the tree to extend. During this it could happen that we discover
a perfectly matching candidate, despite not finding one previously. This may be
because an argument is a container and the container's value gets changed by
another thread. In this case, we may as well take the existing entry and call
it a day.

If instead the argument gets changed so the candidate we add no longer matches,
we're still not off worse, as regardless of how many times we check this, it
may still happen after we dispatch. In such cases it's up to the HLL code to
ensure, that it can cope with this (e.g. locking, knowing that it can only go
from undefined to defined, etc.)

Fixes GH #1237