Remove unneeded special case in MVM_frame_find_lexical_by_name
MVM_frame_find_lexical_by_name throws an exception if the requested lexical
could not be found. However this did not happen for object type lexicals.
According to a comment, this was expected by various things but apparently
it was depended on only in a single place in Perl6::Actions and that was just
a bug. With that fixed, we can get rid of the special handling and some
branching.