Add a probe for pow() corner cases with Inf and NaN.
Solaris doesn't default to being POSIX conformat, but NQP and the spectests
assume that we are. It's not clear *how* to flip the defaults with gcc, so
code around the problem.

Disable JITting `pow_n` if the C library pow() isn't conformant.
(I don't have x86 Solaris to test, but given how thorough Sun's engineering
was, I assume that it's C library behaves identically to Sparc's, so likely
it would JIT to the LTA implementation.)