On AIX, use $config{ptr_size} to conditionally add 32 bit linker flags.
This eliminates the only use of $config{arch_bits} and hence the only use of
build::probe::numbits().

This probe still isn't quite correct - it messes up in some combinations of
flags and compilers, but as we can't currently build "out of the box" on
AIX, defer changing this until we attempt a comprehensive fix.