Add a probe for asin() and acos() returning NaN for out of range input.
Similar to commit 5e031c917587f8b3 for pow(), Solaris also seems to show
non-POSIX behaviour for asin() and acos(), for the same compilers...

As with the others can *sort of* read the behaviour as conformat:
    For finite values of x not in the range [-1,1], a domain error shall
    occur, and
        either a NaN (if supported)
        or an implementation-defined value
    shall be returned.

because here the "implementation-defined value" returned is 0.0
*but* it's not clear of one is allowed to make that choice when one supports
NaNs.

(Maybe yes, because it's "if supported" not "when supported", but it isn't a
totally helpful choice, Was there ever a point when SunOS hardware didn't
support NaNs? Pre-sparc? but it seems that even the Motorola 68881 and 68882
FPUs supported IEEE representations)