Convert MVM_num_{isnanorinf,posinf,neginf,nan} to inline functions.
They are all one-liners - likely the size of the inlined code is smaller
than the calling convention setup.

(Unless one is building with Link Time Optimisation, in which case *that*
will do the transformation, but even then, why make more work for ourselves?)

This has the side effect of moving `#include <math.h>` to a header file,
meaning that various C files no longer need to include it.