Helps build from scratch
After trying to build it, I realized that `make` does not copy to `prefix/bin`, so that's included. Also, this might be unnecessary, but it would be good to clarify that it's not a prefix for the executable, but for everything.
Also, not totally clear make install is actually working. Working on it right now.