Remove the `--instrument` and `--no-instrument` flags from Configure.pl
These would:
    Toggle extra instrumentation flags during compile and link; for example,
    turns on Address Sanitizer when compiling with C<clang>.  Defaults to off.

Except that we now have an explicit --asan option to do this too.

Instead of ASAN, these options (would) actually turn on profiling for gcc,
icc and msvc, but this wasn't documented. Anyone wishing to do this could
also do it by setting the environment variables `CFLAGS` and `LDFLAGS` to
set the relevant options, which is an approach which *is* documented.

Configure.pl is too complex already - we don't need to offer multiple ways to
enable the same infrequently used functionality, particularly for
functionality that is intended for "power" users (in this case developers),
where likely we can only offer a single pre-canned solution for something
which isn't suited to a "one size fits all" use case, let alone this option
which turned into "lucky dip".