Properly clean up profiling data with --full-cleanup
When running with --full-cleanup, we need to free the profiler's call graph
data. However, doing so has revealed, that we have always violated the
profiler's assumption that after starting profiling, we only enter one call
tree. Instead in addition to the call tree of the profiled program, we entered
new call trees for processing END phasers.

To restore the profiler's assumption and greatly ease memory management, we now
do a manual enter/exit when starting/ending the profiling, thus creating an
artificial call frame as parent of the call trees. This works well as profiler
frontends shared the assumption of having one top most call frame. As a side
effect, these now show data about processing of END blocks, which may be
interesting in itself.