Fix leak of MVMUnwindData when a frame exit handler throws
If we're unwinding the stack after an exception and one of the frames has an
unwind handler that itself throws, like in `try { POST { 0 }; die; }`, we
ended up leaking the MVMUnwindData allocated for resuming the initial unwind
after running the handler. Since that handler will itself cause an unwind, we
don't need to continue the initial unwind, but we still have to free the
unwind data.