First steps towards resume initialization state
The resume initialization state is the state that a dispatcher indicates
as being needed in order to start a resumption of a dispatch. From the
perspective of the dispatch program, it's another place to obtain some
values. This gets us to the point of handling the simplest case of the
resume initialization state (that is, where it precisely matches the
arguments of the initial dispatch), and even then only when we are in
the record phases of both the initial dispatch and its resumption. There
is enough updated for it to produce a dispatch program that fails due to
an unimplemented op (rather than silently obtaining an argument from the
wrong place).