Fix type mismatch in struct used for FSA_DEBUG
The FSA functions use size_t for the byte size parameter, but the
struct used when FSA_DEBUG is set used MVMuint64. Change the struct to
use size_t and then correct the *printf format strings that use them.
Also just pass the name of the static char array to snprintf, don't need
to precede it with `&`.