Avoid segfaults caused by repossessed native call sites
We deserialize a native call site. Then we call the function which triggers the
actual setup, i.e. resolving the library name, loading the library, generating
caller code. We run the function often enough that spesh decides its worth its
time. While spesh is working however, we deserialize the module again and as
part of the repossession process clear the native call site's body. I.e.
body->entry_point becomes NULL. This causes us to compile a jump to 0.

The real fix for this is to avoid repossession of the native call site in the
first place. In case the HLL layer neglects to to the appropriate steps, do
what we can to detect the situation and downgrade the segfault to just a missed
optimization opportunity.