Restructure resumption recording
When we will work through a number of possible ways to resume, there is
state and guards involved with each of them. Thus the recording data
structure as we had it so far was not sufficient. This re-organizes it
and does other preparatory work in order to support moving through the
various resumable dispatchers. That doesn't work *yet*; this is just
enough changes to re-work the data structures without regressing
anything (according to the NQP-based test suite for the dispatch
mechanism, anyway).