Always log the type coming out of an nqp::decont
Before, it would only log if a decont was actually performed. That was
usually fine because most of the time there always was or wasn't a
container. If there was a container we logged the type, and if there
wasn't we usually could throw out the decont op altogether. However, in
the cases where there *was* a mix, the stats could be misleading. Given
that we will soon have the ability to remove an optimized spesh candidate
if it gets too many deopts (and then clear the stats), we'll still end up
creating a new candidate just like the previous bad one because its new
stats will be just as misleading as the old ones. With this change we
collect good stats in any case.