Work around libuv behavior change on multiple active reads on async socket
Before the recent libuv bump, libuv would on UNIX ignore an already active read
on an async socket. This caused us to silently replace the existing reader with
the new one. On Windows however, it would return UV_EALREADY causing us to
throw an exception when attaching the second reader. If the exception got
caught we would even run into a panic. With the harmonization of error handling
in libuv commit 06b731742208c0c7b0b56948da2f2e0c9e645ec9 the exception and
panic now also happens on UNIX.

This workaround restores the previous latest wins semantics to buy us time to
figure out how we really want to react to this situation.

Can be reproduced with:
react {
    whenever IO::Socket::Async.listen("localhost", 1234) -> $conn {
        $conn.Supply(:bin).tap: -> $v { say $v }, quit => { say "quit" } for ^2
    }
}
and a client like netcat localhost 1234 sending anything.