Eliminate the ops sec_n, asec_n and sech_h, which are no longer used.
These hyperbolic functions are not provided directly by libc, so have to be
implemented in terms of functions which are. There's no mathematical
difference in expressing the same identities in the Rakudo setting (it's
going to dispatched as the same IEEE operations) but doing it there instead
of here means that the code is easier for spesh and the JIT to optimise.

Rakudo stopped using these ops in the 2021.02 release, NQP in 2021.03.
Now now remove them from MoarVM.