No longer require an exception as argument of nqp::backtrace
Previously generating a backtrace required throwing and catching an exception
first. The exception was just used as a position marker to start the backtrace
from. In many cases however, the nqp::backtrace call itself is perfectly fine
as a starting point so requiring an exception made nqp::backtrace needlessly
expensive.

Furthermore we currently cannot create consistent backtraces at all if the
exception is thrown in a callee of the frame containing the nqp::backtrace op.
The exception object holds a pointer to the frame that it was thrown from.
With this frame one can easily follow the caller chain. However the information
of where exactly in the caller the callee was called from is not retained (at
least for JIT compiled code).

Apart from getting line numbers wrong, this makes it impossible to determine
whether the caller was inlined and thus does not actually exist as its own
call frame. Thus the backtrace may be missing such frames.

Turns out that this happens exactly to one of the frames that would be omitted
from the HLL backtrace anyway as they were part of generating the backtrace in
the first place and not really interesting to the user. The need for consistent
backtraces in the face of inlining was created solely for being able to know
how may frames to skip.

Removing the need for the exception and the call frames that were involved in
generating and catching this exception makes creating Failures ~ 25 % faster
and avoids the inconsistencies caused by inlining.