Fix spesh removing not-really-dead code
For various optimizations, spesh needs an accurate model of the code's control
flow. For example with a hypothetical function like this:

BB 1:
    set r1, lit(0)
BB 2:
    set r1, lit(1)
BB 3:
    return_i r1

the control flow is obviously BB(1) -> BB(2) -> BB(3). BB(2) is the one and
only successor of BB(1) and BB(3) is the one and only successor of BB(2).

The first set operation is therefore quite unnecessary and can be removed
without harm as control flow will inevitable move on to the second set
operation which would overwrite the result of the first op right away. Since
set has no side-effects and its result is not used, we can simply remove it.
The function will always return 1 either way.

To do so, spesh needs to prove that control flow will end up in a place that
overwrites the first set's result before the result register is read by
anything. It creates a complete control flow graph taking branches and jumps
into account. For example in:

BB 1:
    set r1, lit(0)
    unless_i r1, BB(3)
BB 2:
    set r1, lit(1)
BB 3:
    return_i r1

the control flow graph splits:
BB(1)
 |  \
 v   \
BB(2) |
 |   /
 v  v
BB(3)

BB(1) now has 2 possible successors: BB(2) and BB(3), while BB(2) still has
only one way to go and only BB(3) as successor.

It's no longer a given that the first set's result is going to be overwritten.
Indeed, since we know that result (a static 0) and the result of the unless_i
comparison, we actually know that it's the second set op that is useless, as it
is never run at all.

Things become more difficult, when we need to take exceptions into account.
Exceptions and exception handlers form yet another way to influence control
flow:

BB 1:
    set r1, lit(0)
    [Annotation: FH Start (0)] # start of the area covered by the exception handler
    invoke_o r2, r3 # can throw an exception
BB 2:
    [Annotation: FH End (0)] # end of the handler's area
    set r1, lit(1)
    goto BB(4)
BB 3:
    [Annotation: FH Goto (0)] # exception handler
    takehandlerresult r4
BB 4:
    return_i r1

This time the control flow graph turns out to be:
BB(1)
 |   \
 |    \
 v     v
BB(2) BB(3)
 |    /
 |   /
 v  v
BB(4)

So BB(1) again has 2 successors: BB(2) and BB(3) while those two each have
BB(4) as only successor.

There are 2 types of exception: control exceptions and catch exceptions. For
control exceptions, spesh was already modeling control flow correclty. However
for unknown reasons, they were special cased in the code and catch exceptions
were not handled similarily. This resulted in the edge from BB(1) to BB(3)
missing. Thus to spesh it looked like we would always reach BB(2) (as it's the
only possible successor of BB(1)) thus the first set's result would always be
overwritten.

That was not the case with with Perl6::World's is_type method:

    method is_type(@name) {
        my $is_name := 0;
        try {
            # This throws if it's not a known name.
            $is_name := !nqp::isconcrete(self.find_symbol(@name))
        }
        $is_name
    }

Since spesh thought it could eliminate the initialization of $is_name, if
the symbol was not found, $is_name was left uninitialized. Since the function
must return an object, this meant that instead of a boxed 0, it would write
NULL into the return register. Since NULL must never be used as an object
register, this could trip up some later operation.

Fix by simply removing the case distinction. The code for control exceptions
is already perfectly applicable to catch exceptions.